+++ www/go/core/views/extjs3/detail/Panel.js
@@ -204,6 +204,10 @@ go.detail.Panel = Ext.extend(Ext.Panel, {
 		
 	},
 
+	print() {
+		this.body.print();
+	},
+
 	load: function (id) {
 
 		id = parseInt(id);